requirements.txt
torchmetrics
einops
pytest
matplotlib
